requirements.txt
flask
flask-sqlalchemy
flask-migrate
gunicorn
psycopg2-binary
python-dotenv
dropbox
